ltx: extract per-model lifecycle adapters out of the pipeline
Move config parsing, module construction, and disk-cache weight loading
out of LTXPipeline and into per-model owners, so the pipeline holds thin
delegators instead of driving every load path itself:

- LTXVideoVAEAdapter (vae_ltx.py) owns the video VAE decoder + encoder,
  their shared per-channel stats, and reload_decoder/reload_encoder.
- LTXAudioDecoderAdapter (audio_decoder_ltx.py) owns the mel-VAE decoder
  + vocoder/BWE stack, audio parallel-config selection, and reload.
- LTXTransformerCheckpoint (transformer_ltx.py) owns variant detection,
  LoRA-tagged cache naming, state-dict fusing, build, and load.
- LTXLatentUpsampler.from_checkpoint/reload_weights owns its own config
  read and cache load.

Split the mel-VAE decoder (MelDecoder, renamed from AudioDecoder) into
mel_decoder_ltx.py, leaving audio_decoder_ltx.py as the lifecycle adapter.

TransformerState now carries the checkpoint + lora_specs instead of a
cache name and state-dict provider closure. vae_decoder/vae_encoder/
tt_mel_decoder/tt_vocoder_with_bwe become read-through properties.

Drop the SUPPORTS_IMAGE_CONDITIONING class flag: I2V is enabled whenever
the checkpoint's VAE has encoder blocks. Update audio/distilled tests to
construct the adapter directly.
