+++ models/tt_dit/models/audio_vae/audio_decoder_ltx.py
@@ -2,521 +2,267 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-"""LTX-2 audio mel-VAE decoder (Stage A).
+"""Lifecycle adapter for the LTX-2 audio decode stack (mel-VAE decoder + vocoder/BWE).
 
-Conv2dViaConv3d operates on ``(B, H, W, C)`` (mel_bins is W); the torch reference
-is BCHW, so inputs convert at the device boundary and back on exit. All convs are
-causal on the height (time) axis. Single-chip.
+Owns config parsing, module construction (including the audio parallel-config selection), and the
+disk-cache weight load — mirroring ``LTXVideoVAEAdapter`` for video. There is no WAN analog (WAN
+has no audio), but the shape is identical: the pipeline holds an adapter and calls thin delegators.
 """
 
 from __future__ import annotations
 
-from collections import OrderedDict
+import json
+import os
 
-import einops
 import torch
+from loguru import logger
+from safetensors import safe_open
 
 import ttnn
 
-from ...layers.audio_ops import Conv2dViaConv3d
-from ...layers.module import Module, ModuleList
-from ...utils.conv3d import conv_pad_in_channels
-from ...utils.tracing import Tracer
+from ...parallel.config import AudioTCParallelConfig, DiTParallelConfig, ParallelFactor
+from ...parallel.manager import CCLManager
+from ...utils import cache as cache_module
+from ...utils.conv3d import conv3d_blocking_hash
+from .bwe_ltx import MelSTFT, VocoderWithBWE
+from .mel_decoder_ltx import MelDecoder
+from .vocoder_ltx import Vocoder
 
-LATENT_DOWNSAMPLE_FACTOR = 4
-
-
-class AudioPatchifier:
-    """Host-side patchifier with ``patch_size=1``: flatten/unflatten over (channels, mel_bins)."""
-
-    def patchify(self, audio_latents: torch.Tensor) -> torch.Tensor:
-        return einops.rearrange(audio_latents, "b c t f -> b t (c f)")
-
-    def unpatchify(self, audio_latents: torch.Tensor, channels: int, mel_bins: int) -> torch.Tensor:
-        return einops.rearrange(audio_latents, "b t (c f) -> b c t f", c=channels, f=mel_bins)
-
-
-class PixelNorm(Module):
-    """Per-pixel RMS normalization over the channel (last) axis. No learned params."""
-
-    def __init__(self, eps: float = 1e-6) -> None:
-        super().__init__()
-        self.eps = eps
-
-    def forward(self, x_BHWC: ttnn.Tensor) -> ttnn.Tensor:
-        x_sq = ttnn.multiply(x_BHWC, x_BHWC)
-        mean_sq = ttnn.mean(x_sq, dim=-1, keepdim=True)
-        rms = ttnn.sqrt(ttnn.add(mean_sq, self.eps))
-        return ttnn.multiply(x_BHWC, ttnn.reciprocal(rms))
+# Config keys forwarded verbatim from the checkpoint's vocoder/bwe config into ``Vocoder``.
+_VOCODER_CONFIG_KEYS = (
+    "resblock_kernel_sizes",
+    "upsample_rates",
+    "upsample_kernel_sizes",
+    "resblock_dilation_sizes",
+    "upsample_initial_channel",
+    "resblock",
+    "activation",
+    "use_tanh_at_final",
+    "use_bias_at_final",
+)
 
 
-class ResnetBlock(Module):
-    """LTX-2 audio mel-VAE Resnet block. No temb branch, no dropout."""
+class LTXAudioDecoderAdapter:
+    """Owns the LTX audio decode stack: parses the checkpoint's ``audio_vae`` / ``vocoder`` config,
+    builds the ``MelDecoder`` + ``VocoderWithBWE`` (two vocoders via the BWE path), selects the
+    audio parallel config, and loads/reloads their weights.
 
-    def __init__(
-        self,
-        *,
-        in_channels: int,
-        out_channels: int | None = None,
-        mesh_device: ttnn.MeshDevice,
-        dtype: ttnn.DataType = ttnn.bfloat16,
-    ) -> None:
-        super().__init__()
-
-        out_channels = in_channels if out_channels is None else out_channels
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-        self.has_shortcut = in_channels != out_channels
-
-        self.norm1 = PixelNorm()
-        self.conv1 = Conv2dViaConv3d(
-            in_channels,
-            out_channels,
-            kernel_size=3,
-            stride=1,
-            padding_mode="causal_height",
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-        self.norm2 = PixelNorm()
-        self.conv2 = Conv2dViaConv3d(
-            out_channels,
-            out_channels,
-            kernel_size=3,
-            stride=1,
-            padding_mode="causal_height",
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-        if self.has_shortcut:
-            self.nin_shortcut = Conv2dViaConv3d(
-                in_channels,
-                out_channels,
-                kernel_size=1,
-                stride=1,
-                padding_mode="causal_height",
-                mesh_device=mesh_device,
-                dtype=dtype,
-            )
-
-    def _prepare_torch_state(self, state: dict[str, torch.Tensor]) -> None:
-        keys_to_remove = [k for k in state if k.startswith("non_linearity") or k.startswith("dropout")]
-        for k in keys_to_remove:
-            del state[k]
-
-        if not self.has_shortcut:
-            keys_to_remove = [k for k in state if k.startswith("nin_shortcut") or k.startswith("conv_shortcut")]
-            for k in keys_to_remove:
-                del state[k]
-
-    def forward(self, x_BHWC: ttnn.Tensor) -> ttnn.Tensor:
-        residual = x_BHWC
-
-        h = self.norm1(x_BHWC)
-        h = ttnn.silu(h)
-        h = ttnn.to_layout(h, ttnn.ROW_MAJOR_LAYOUT) if h.layout != ttnn.ROW_MAJOR_LAYOUT else h
-        h = self.conv1(h)
-
-        h = self.norm2(h)
-        h = ttnn.silu(h)
-        h = ttnn.to_layout(h, ttnn.ROW_MAJOR_LAYOUT) if h.layout != ttnn.ROW_MAJOR_LAYOUT else h
-        h = self.conv2(h)
-
-        if self.has_shortcut:
-            residual = (
-                ttnn.to_layout(residual, ttnn.ROW_MAJOR_LAYOUT)
-                if residual.layout != ttnn.ROW_MAJOR_LAYOUT
-                else residual
-            )
-            residual = self.nin_shortcut(residual)
-
-        return ttnn.add(residual, h)
-
-
-class AudioUpsample(Module):
-    """Nearest-neighbour 2x upsample + causal conv + drop leading H row."""
-
-    def __init__(
-        self,
-        *,
-        in_channels: int,
-        mesh_device: ttnn.MeshDevice,
-        dtype: ttnn.DataType = ttnn.bfloat16,
-    ) -> None:
-        super().__init__()
-        self.in_channels = in_channels
-        self.conv = Conv2dViaConv3d(
-            in_channels,
-            in_channels,
-            kernel_size=3,
-            stride=1,
-            padding_mode="causal_height",
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-
-    def forward(self, x_BHWC: ttnn.Tensor) -> ttnn.Tensor:
-        # ttnn.upsample requires ROW_MAJOR input and produces ROW_MAJOR output.
-        x_BHWC = ttnn.to_layout(x_BHWC, ttnn.ROW_MAJOR_LAYOUT) if x_BHWC.layout != ttnn.ROW_MAJOR_LAYOUT else x_BHWC
-        x_BHWC = ttnn.upsample(x_BHWC, scale_factor=2)
-        x_BHWC = self.conv(x_BHWC)
-
-        # Drop the leading H row to undo the causal pad: output length is 1 + 2*n_in.
-        x_BHWC = x_BHWC[:, 1:, :, :]
-        return x_BHWC
-
-
-class _MidBlock(Module):
-    """Mid block: two Resnet blocks, no attention."""
-
-    def __init__(
-        self,
-        *,
-        channels: int,
-        mesh_device: ttnn.MeshDevice,
-        dtype: ttnn.DataType = ttnn.bfloat16,
-    ) -> None:
-        super().__init__()
-        self.block_1 = ResnetBlock(
-            in_channels=channels,
-            out_channels=channels,
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-        self.block_2 = ResnetBlock(
-            in_channels=channels,
-            out_channels=channels,
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-
-    def _prepare_torch_state(self, state: dict[str, torch.Tensor]) -> None:
-        keys_to_remove = [k for k in state if k.startswith("attn_1")]
-        for k in keys_to_remove:
-            del state[k]
-
-    def forward(self, x_BHWC: ttnn.Tensor) -> ttnn.Tensor:
-        x_BHWC = self.block_1(x_BHWC)
-        x_BHWC = self.block_2(x_BHWC)
-        return x_BHWC
-
-
-class _UpStage(Module):
-    """One upsampling level: ``block`` (ModuleList) + optional ``upsample``."""
-
-    def __init__(
-        self,
-        *,
-        in_channels: int,
-        out_channels: int,
-        num_res_blocks_in_stage: int,
-        has_upsample: bool,
-        mesh_device: ttnn.MeshDevice,
-        dtype: ttnn.DataType = ttnn.bfloat16,
-    ) -> None:
-        super().__init__()
-
-        # First block does the channel transition; the rest are out → out.
-        self.block = ModuleList()
-        cur_in = in_channels
-        for _ in range(num_res_blocks_in_stage):
-            self.block.append(
-                ResnetBlock(
-                    in_channels=cur_in,
-                    out_channels=out_channels,
-                    mesh_device=mesh_device,
-                    dtype=dtype,
-                )
-            )
-            cur_in = out_channels
-
-        self.has_upsample = has_upsample
-        if has_upsample:
-            self.upsample = AudioUpsample(
-                in_channels=out_channels,
-                mesh_device=mesh_device,
-                dtype=dtype,
-            )
-
-    def _prepare_torch_state(self, state: dict[str, torch.Tensor]) -> None:
-        keys_to_remove = [k for k in state if k.startswith("attn.")]
-        for k in keys_to_remove:
-            del state[k]
-        if not self.has_upsample:
-            keys_to_remove = [k for k in state if k.startswith("upsample.")]
-            for k in keys_to_remove:
-                del state[k]
-
-    def forward(self, x_BHWC: ttnn.Tensor) -> ttnn.Tensor:
-        for block in self.block:
-            x_BHWC = block(x_BHWC)
-        if self.has_upsample:
-            x_BHWC = self.upsample(x_BHWC)
-        return x_BHWC
-
-
-class AudioDecoder(Module):
-    """LTX-2 audio mel-VAE decoder.
-
-    ``forward`` accepts a torch tensor ``(B, z_channels, frames, mel_bins)`` and
-    returns ``(B, out_ch, T_target, mel_bins)``.
+    ``mel_decoder`` / ``vocoder_with_bwe`` expose the underlying modules (both ``None`` when the
+    checkpoint requires an unsupported norm/causality). ``reload_weights`` is idempotent and driven
+    by the pipeline; it re-injects the mel-VAE per-channel stats after load (load-bearing — those
+    buffers are not carried by the binary cache).
     """
 
     def __init__(
         self,
+        checkpoint_path: str,
         *,
-        ch: int,
-        out_ch: int,
-        ch_mult: tuple[int, ...] = (1, 2, 4),
-        num_res_blocks: int,
-        attn_resolutions: set[int] | tuple[int, ...] = (),
-        resolution: int,
-        z_channels: int,
-        mid_block_add_attention: bool = False,
-        sample_rate: int = 16000,
-        mel_hop_length: int = 160,
-        is_causal: bool = True,
-        mel_bins: int | None = None,
         mesh_device: ttnn.MeshDevice,
-        dtype: ttnn.DataType = ttnn.bfloat16,
+        vae_ccl_manager: CCLManager,
+        dit_parallel_config: DiTParallelConfig,
+        traced: bool,
     ) -> None:
-        super().__init__()
+        self._checkpoint_path = checkpoint_path
+        self._mesh_device = mesh_device
+        self._vae_ccl_manager = vae_ccl_manager
+        self._dit_parallel_config = dit_parallel_config
+        self._traced = traced
 
-        if attn_resolutions:
-            raise NotImplementedError(
-                "AudioDecoder Stage A does not support attention blocks; " f"got attn_resolutions={attn_resolutions}"
-            )
-        if mid_block_add_attention:
-            raise NotImplementedError("AudioDecoder Stage A does not support mid_block_add_attention=True")
+        self._mel_decoder: MelDecoder | None = None
+        self._vocoder_with_bwe: VocoderWithBWE | None = None
 
-        self.out_ch = out_ch
-        self.ch_mult = tuple(ch_mult)
-        self.num_res_blocks = num_res_blocks
-        self.num_resolutions = len(self.ch_mult)
-        self.z_channels = z_channels
-        self.is_causal = is_causal
-        self.mel_bins = mel_bins
-        self.mesh_device = mesh_device
+        self._build()
 
-        # Per-channel denormalize stats: non-Parameter host tensors consumed in
-        # _denormalize_latents. Default to identity (std=1, mean=0) so an unset
-        # path is a harmless passthrough rather than uninitialized garbage.
-        self._stats_std = torch.ones(ch)
-        self._stats_mean = torch.zeros(ch)
+    @property
+    def mel_decoder(self) -> "MelDecoder | None":
+        return self._mel_decoder
 
-        # forward_traced state, mirroring Vocoder: capture the pure-device graph once per
-        # input shape and replay it (the conv ops are host-dispatch-bound, so this removes
-        # the dominant per-op dispatch cost). _target_shape is set per-input by
-        # _host_to_device and read by _device_to_host, so it is not baked into the graph.
-        self.use_trace = False
-        self._target_shape: tuple[int, int, int, int] | None = None
-        self._traces: "OrderedDict[tuple, Tracer]" = OrderedDict()
-        self._max_traces = None
+    @property
+    def vocoder_with_bwe(self) -> "VocoderWithBWE | None":
+        return self._vocoder_with_bwe
 
-        self.patchifier = AudioPatchifier()
+    def _build(self) -> None:
+        """Construct audio decoder + vocoder module shells from checkpoint config.
 
-        base_block_channels = ch * self.ch_mult[-1]
-
-        self.conv_in = Conv2dViaConv3d(
-            z_channels,
-            base_block_channels,
-            kernel_size=3,
-            stride=1,
-            padding_mode="causal_height",
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-
-        self.mid = _MidBlock(
-            channels=base_block_channels,
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-
-        # Build high→low resolution so channel counts propagate, storing by level;
-        # then append in level order so state-dict keys ``up.<level>.*`` match
-        # (forward iterates these in reverse).
-        stages: list[_UpStage] = [None] * self.num_resolutions
-        block_in = base_block_channels
-        for level in reversed(range(self.num_resolutions)):
-            block_out = ch * self.ch_mult[level]
-            stages[level] = _UpStage(
-                in_channels=block_in,
-                out_channels=block_out,
-                num_res_blocks_in_stage=self.num_res_blocks + 1,
-                has_upsample=level != 0,
-                mesh_device=mesh_device,
-                dtype=dtype,
-            )
-            block_in = block_out
-        self.up = ModuleList()
-        for stage in stages:
-            self.up.append(stage)
-
-        final_block_channels = block_in
-
-        self.norm_out = PixelNorm()
-        self.conv_out = Conv2dViaConv3d(
-            final_block_channels,
-            out_ch,
-            kernel_size=3,
-            stride=1,
-            padding_mode="causal_height",
-            mesh_device=mesh_device,
-            dtype=dtype,
-        )
-
-    def _prepare_torch_state(self, state: dict[str, torch.Tensor]) -> None:
-        # per_channel_statistics buffers carry hyphens (not valid Python idents),
-        # so consume them on host rather than registering them as children.
-        std_key = "per_channel_statistics.std-of-means"
-        mean_key = "per_channel_statistics.mean-of-means"
-        if std_key in state:
-            self._stats_std = state.pop(std_key).detach().clone()
-        if mean_key in state:
-            self._stats_mean = state.pop(mean_key).detach().clone()
-
-        keys_to_remove = [k for k in state if k.startswith("non_linearity")]
-        for k in keys_to_remove:
-            del state[k]
-
-    def set_per_channel_stats(self, std: torch.Tensor, mean: torch.Tensor) -> None:
-        """Re-inject the per-channel denormalize stats after a binary cache load.
-
-        These are non-Parameter host buffers; the weight cache carries no
-        non-Parameter state, so without re-injection the denormalize multiplies
-        by uninitialised memory.
+        No weights are loaded — ``reload_weights`` handles that via the disk cache the same way
+        the video VAE adapter does.
         """
-        self._stats_std = std.detach().clone().float()
-        self._stats_mean = mean.detach().clone().float()
+        with safe_open(self._checkpoint_path, framework="pt") as f:
+            config = json.loads(f.metadata()["config"])
 
-    def _denormalize_latents(self, sample: torch.Tensor) -> tuple[torch.Tensor, tuple[int, int, int, int]]:
-        """Apply ``un_normalize`` on the patchified latents, then unpatchify.
+        ad = config["audio_vae"]["model"]["params"]
+        ddconfig = ad["ddconfig"]
+        stft_cfg = config["audio_vae"].get("preprocessing", {}).get("stft", {})
+        mel_cfg = config["audio_vae"].get("preprocessing", {}).get("mel", {})
+        mel_bins = ddconfig.get("mel_bins") or mel_cfg.get("n_mel_channels")
+        if ddconfig.get("norm_type", "pixel") != "pixel" or ddconfig.get("causality_axis", "height") != "height":
+            logger.warning("Audio decoder: checkpoint requires unsupported norm/causality; skipping construction")
+            return
 
-        Returns the denormalized BCTF tensor *and* the target output shape
-        ``(B, out_ch, target_frames, target_mel_bins)``.
-        """
-        B, C, T, F = sample.shape
+        voc_cfg = config["vocoder"]["vocoder"]
+        bwe_cfg = config["vocoder"]["bwe"]
 
-        patched = self.patchifier.patchify(sample)
-
-        # Stats length depends on the checkpoint: either C*F (direct broadcast)
-        # or C (broadcast over mel_bins via repeat_interleave).
-        std = self._stats_std.to(patched.dtype).to(patched.device)
-        mean = self._stats_mean.to(patched.dtype).to(patched.device)
-        last_dim = patched.shape[-1]
-        if std.shape[0] == last_dim:
-            pass
-        elif std.shape[0] == C and last_dim == C * F:
-            std = std.repeat_interleave(F)
-            mean = mean.repeat_interleave(F)
+        mesh_shape = tuple(self._mesh_device.shape)
+        t_axis = 0 if mesh_shape[0] >= mesh_shape[1] else 1
+        t_factor = mesh_shape[t_axis]
+        c_axis = 1 - t_axis
+        c_factor = mesh_shape[c_axis]
+        channel_tp_on = os.environ.get("LTX_AUDIO_CHANNEL_TP", "0") == "1"
+        if t_factor > 1 and c_factor > 1 and channel_tp_on:
+            audio_parallel_config = AudioTCParallelConfig(
+                time_parallel=ParallelFactor(factor=t_factor, mesh_axis=t_axis),
+                channel_parallel=ParallelFactor(factor=c_factor, mesh_axis=c_axis),
+            )
+        elif t_factor > 1:
+            audio_parallel_config = ParallelFactor(factor=t_factor, mesh_axis=t_axis)
         else:
-            raise ValueError(
-                f"per-channel stats shape {tuple(std.shape)} does not broadcast against "
-                f"patched last dim {last_dim} (C={C}, F={F})"
-            )
-        sample_denormalized = patched * std + mean
+            audio_parallel_config = None
+        audio_ccl = self._vae_ccl_manager if audio_parallel_config is not None else None
 
-        sample = self.patchifier.unpatchify(sample_denormalized, channels=C, mel_bins=F)
-
-        target_frames = T * LATENT_DOWNSAMPLE_FACTOR
-        if self.is_causal:
-            target_frames = max(target_frames - (LATENT_DOWNSAMPLE_FACTOR - 1), 1)
-        target_mel_bins = self.mel_bins if self.mel_bins is not None else F
-        target_shape = (B, self.out_ch, target_frames, target_mel_bins)
-
-        return sample, target_shape
-
-    @staticmethod
-    def _adjust_output_shape(decoded_output: torch.Tensor, target_shape: tuple[int, int, int, int]) -> torch.Tensor:
-        """Crop / pad the decoded output to the exact target shape."""
-        _, _, current_time, current_freq = decoded_output.shape
-        _, target_channels, target_time, target_freq = target_shape
-
-        decoded_output = decoded_output[
-            :, :target_channels, : min(current_time, target_time), : min(current_freq, target_freq)
-        ]
-
-        time_padding_needed = target_time - decoded_output.shape[2]
-        freq_padding_needed = target_freq - decoded_output.shape[3]
-        if time_padding_needed > 0 or freq_padding_needed > 0:
-            padding = (
-                0,
-                max(freq_padding_needed, 0),
-                0,
-                max(time_padding_needed, 0),
-            )
-            decoded_output = torch.nn.functional.pad(decoded_output, padding)
-
-        return decoded_output[:, :target_channels, :target_time, :target_freq]
-
-    def _host_to_device(self, latent: torch.Tensor) -> ttnn.Tensor:
-        """Host preprocessing + upload. Split out so the pure-device graph can be captured
-        for trace replay; sets ``self._target_shape`` (input-dependent, read by
-        ``_device_to_host``) so it is not baked into the captured graph."""
-        sample_bcfk, target_shape = self._denormalize_latents(latent)
-        self._target_shape = target_shape
-
-        # To BHWC for the device (mel_bins is W).
-        sample_bhwc = sample_bcfk.permute(0, 2, 3, 1).contiguous().to(torch.float32)
-
-        # The input tensor must carry the same padded channel count as the
-        # Conv2dViaConv3d weight.
-        sample_bhwc_padded = conv_pad_in_channels(sample_bhwc)
-
-        return ttnn.from_torch(
-            sample_bhwc_padded, device=self.mesh_device, layout=ttnn.ROW_MAJOR_LAYOUT, dtype=ttnn.bfloat16
+        self._mel_decoder = MelDecoder(
+            ch=ddconfig.get("ch", 128),
+            out_ch=ddconfig.get("out_ch", 2),
+            ch_mult=tuple(ddconfig.get("ch_mult", (1, 2, 4))),
+            num_res_blocks=ddconfig.get("num_res_blocks", 2),
+            attn_resolutions=tuple(ddconfig.get("attn_resolutions", ())),
+            resolution=ddconfig.get("resolution", 256),
+            z_channels=ddconfig.get("z_channels", 8),
+            mid_block_add_attention=ddconfig.get("mid_block_add_attention", False),
+            sample_rate=ad.get("sampling_rate", 16000),
+            mel_hop_length=stft_cfg.get("hop_length", 160),
+            is_causal=stft_cfg.get("causal", True),
+            mel_bins=mel_bins,
+            mesh_device=self._mesh_device,
+            dtype=ttnn.bfloat16,
         )
 
-    def _forward_device(self, x: ttnn.Tensor) -> ttnn.Tensor:
-        """The pure on-device decode graph (conv_in → conv_out). Captured for trace replay."""
-        x = self.conv_in(x)
-        x = self.mid(x)
+        if isinstance(audio_parallel_config, AudioTCParallelConfig):
+            override = os.environ.get("LTX_BWE_CHANNEL_TP")
+            use_bwe_ctp = override == "1" if override is not None else audio_parallel_config.channel_parallel.factor > 2
+            bwe_pc = audio_parallel_config if use_bwe_ctp else audio_parallel_config.time_parallel
+        else:
+            bwe_pc = audio_parallel_config
+        main_voc = self._build_vocoder(
+            voc_cfg, apply_final_activation=True, parallel_config=audio_parallel_config, ccl_manager=audio_ccl
+        )
+        bwe_voc = self._build_vocoder(
+            bwe_cfg, apply_final_activation=False, parallel_config=bwe_pc, ccl_manager=audio_ccl
+        )
+        mel_stft = MelSTFT(
+            filter_length=bwe_cfg["n_fft"],
+            hop_length=bwe_cfg["hop_length"],
+            win_length=bwe_cfg["n_fft"],
+            n_mel_channels=bwe_cfg["num_mels"],
+            mesh_device=self._mesh_device,
+            dtype=ttnn.float32,
+        )
+        self._vocoder_with_bwe = VocoderWithBWE(
+            vocoder=main_voc,
+            bwe_generator=bwe_voc,
+            mel_stft=mel_stft,
+            input_sampling_rate=bwe_cfg["input_sampling_rate"],
+            output_sampling_rate=bwe_cfg["output_sampling_rate"],
+            hop_length=bwe_cfg["hop_length"],
+            mesh_device=self._mesh_device,
+            dtype=ttnn.float32,
+        )
+        self._vocoder_with_bwe.use_trace = self._traced and os.environ.get("LTX_VOC_TRACE", "1") != "0"
+        self._vocoder_with_bwe.use_trace_bwe = self._traced
+        self._mel_decoder.use_trace = self._traced and os.environ.get("LTX_VAE_TRACE", "0") == "1"
+        if isinstance(audio_parallel_config, AudioTCParallelConfig):
+            cfg_desc = f"T-shard={t_factor} axis{t_axis} + channel-TP={c_factor} axis{c_axis}"
+        elif audio_parallel_config is not None:
+            cfg_desc = f"T-shard={t_factor} axis{t_axis} (single-axis)"
+        else:
+            cfg_desc = "replicated"
+        logger.info(f"Constructed audio decoder shells (mesh {mesh_shape}, vocoder {cfg_desc})")
 
-        for level in reversed(range(self.num_resolutions)):
-            stage = self.up[level]
-            x = stage(x)
+    def _build_vocoder(
+        self,
+        cfg: dict,
+        *,
+        apply_final_activation: bool,
+        parallel_config,
+        ccl_manager,
+    ) -> Vocoder:
+        """Construct a ``Vocoder`` from a checkpoint vocoder/bwe config block.
 
-        x = self.norm_out(x)
-        x = ttnn.silu(x)
-        x = ttnn.to_layout(x, ttnn.ROW_MAJOR_LAYOUT) if x.layout != ttnn.ROW_MAJOR_LAYOUT else x
-        return self.conv_out(x)
-
-    def _device_to_host(self, x: ttnn.Tensor) -> torch.Tensor:
-        """Download + strip out_channels padding (Conv2dViaConv3d) + BHWC → BCHW + reshape."""
-        out_bhwc = ttnn.to_torch(ttnn.get_device_tensors(x)[0])
-        out_bhwc = out_bhwc[..., : self.out_ch]
-        out_bchw = out_bhwc.permute(0, 3, 1, 2).contiguous()
-        return self._adjust_output_shape(out_bchw, self._target_shape)
-
-    def forward(self, latent: torch.Tensor) -> torch.Tensor:
-        """``latent``: ``(B, z_channels, frames, mel_bins)`` →
-        ``(B, out_ch, target_frames, target_mel_bins)``.
+        Used for both the main vocoder and the BWE generator, which differ only in their config
+        block, final-activation behavior, and parallel config.
         """
-        x = self._host_to_device(latent)
-        y = self._forward_device(x)
-        return self._device_to_host(y)
+        return Vocoder(
+            **{k: cfg[k] for k in _VOCODER_CONFIG_KEYS if k in cfg},
+            apply_final_activation=apply_final_activation,
+            mesh_device=self._mesh_device,
+            dtype=ttnn.float32,
+            parallel_config=parallel_config,
+            ccl_manager=ccl_manager,
+        )
 
-    def forward_traced(self, latent: torch.Tensor) -> torch.Tensor:
-        """Same result as ``forward`` but captures the device graph once per input shape and
-        replays it, removing per-op host dispatch. Requires the mesh opened with a
-        ``trace_region_size`` large enough for the resident trace."""
-        shape_key = tuple(latent.shape)
-        tracer = self._traces.get(shape_key)
-        if tracer is None:
-            tracer = Tracer(self._forward_device, device=self.mesh_device, prep_run=True, clone_prep_inputs=True)
-            self._traces[shape_key] = tracer
-            if self._max_traces is not None:
-                while len(self._traces) > self._max_traces:
-                    _, evicted = self._traces.popitem(last=False)  # LRU
-                    evicted.release_trace()
-        self._traces.move_to_end(shape_key)  # LRU touch
-        # _host_to_device sets self._target_shape for this shape (read by _device_to_host).
-        y = tracer(self._host_to_device(latent), tracer_blocking_execution=False, tracer_execute_on_capture=False)
-        return self._device_to_host(y)
+    def _audio_decoder_state_provider(self) -> dict[str, torch.Tensor]:
+        """Audio mel-VAE decoder weights from the checkpoint, prefix-stripped to the module keys."""
+        logger.info("Audio decoder cache miss — loading from checkpoint")
+        state: dict[str, torch.Tensor] = {}
+        with safe_open(self._checkpoint_path, framework="pt") as f:
+            for k in f.keys():
+                if k.startswith("audio_vae.decoder."):
+                    state[k.removeprefix("audio_vae.decoder.")] = f.get_tensor(k).float()
+                elif k.startswith("audio_vae.per_channel_statistics."):
+                    state[k.removeprefix("audio_vae.")] = f.get_tensor(k).float()
+        return state
 
-    def release_trace(self) -> None:
-        """Free all captured traces (call on shutdown, or before re-warming a new bucket set)."""
-        for tracer in self._traces.values():
-            tracer.release_trace()
-        self._traces.clear()
+    def _vocoder_state_provider(self) -> dict[str, torch.Tensor]:
+        """Vocoder (+ BWE) weights from the checkpoint, prefix-stripped to the module keys."""
+        logger.info("Audio vocoder cache miss — loading from checkpoint")
+        state: dict[str, torch.Tensor] = {}
+        with safe_open(self._checkpoint_path, framework="pt") as f:
+            for k in f.keys():
+                if k.startswith("vocoder."):
+                    state[k.removeprefix("vocoder.")] = f.get_tensor(k).float()
+        return state
+
+    def reload_weights(self) -> None:
+        """Load audio decoder + vocoder weights via the disk cache (mirrors ``reload_decoder``).
+
+        The per-channel denormalize stats are re-injected separately after load: they are
+        non-Parameter buffers, so the binary cache path doesn't carry them and would otherwise
+        leave garbage stats (and a garbage audio track).
+        """
+        if self._mel_decoder is None or self._vocoder_with_bwe is None:
+            return
+        if self._mel_decoder.is_loaded() and self._vocoder_with_bwe.is_loaded():
+            return
+
+        model_name = os.path.basename(self._checkpoint_path).removesuffix(".safetensors")
+        blocking_key = conv3d_blocking_hash(self._vocoder_with_bwe)
+        dec_subfolder = f"audio_dec_{blocking_key}" if blocking_key else "audio_dec"
+        voc_subfolder = f"audio_voc_{blocking_key}" if blocking_key else "audio_voc"
+
+        if not self._mel_decoder.is_loaded():
+            cache_module.load_model(
+                self._mel_decoder,
+                model_name=model_name,
+                subfolder=dec_subfolder,
+                parallel_config=self._dit_parallel_config,
+                mesh_shape=tuple(self._mesh_device.shape),
+                get_torch_state_dict=self._audio_decoder_state_provider,
+            )
+
+        if not self._vocoder_with_bwe.is_loaded():
+            cache_module.load_model(
+                self._vocoder_with_bwe,
+                model_name=model_name,
+                subfolder=voc_subfolder,
+                parallel_config=self._dit_parallel_config,
+                mesh_shape=tuple(self._mesh_device.shape),
+                get_torch_state_dict=self._vocoder_state_provider,
+            )
+
+        # Re-inject the mel-VAE per-channel stats from the checkpoint — see docstring.
+        with safe_open(self._checkpoint_path, framework="pt") as f:
+            self._mel_decoder.set_per_channel_stats(
+                f.get_tensor("audio_vae.per_channel_statistics.std-of-means"),
+                f.get_tensor("audio_vae.per_channel_statistics.mean-of-means"),
+            )
+
+        logger.info("Loaded TTNN audio decoder + vocoder")
