+++ models/tt_dit/tests/models/ltx/test_pipeline_ltx_distilled.py
@@ -17,6 +17,7 @@ import torch
 from loguru import logger
 
 import ttnn
+from models.tt_dit.models.audio_vae.audio_decoder_ltx import LTXAudioDecoderAdapter
 from models.tt_dit.pipelines.ltx.pipeline_ltx_distilled import LTXDistilledPipeline
 from models.tt_dit.utils.ltx import (
     DEFAULT_LTX_PROMPT,
@@ -587,7 +588,13 @@ def test_audio_decode_girl(mesh_device, mesh_shape, sp_axis, tp_axis, num_links,
         width=width,
     )
     pipeline.checkpoint_name = LTXDistilledPipeline._resolve_checkpoint_file(ckpt)
-    pipeline._new_audio_decoder()
+    pipeline._audio_adapter = LTXAudioDecoderAdapter(
+        pipeline.checkpoint_name,
+        mesh_device=pipeline.mesh_device,
+        vae_ccl_manager=pipeline.vae_ccl_manager,
+        dit_parallel_config=pipeline.parallel_config,
+        traced=pipeline._traced,
+    )
     pipeline._prepare_audio_decoder()
 
     vps = VideoPixelShape(batch=1, frames=num_frames, height=height, width=width, fps=24)
@@ -641,9 +648,9 @@ def test_audio_decode_girl(mesh_device, mesh_shape, sp_axis, tp_axis, num_links,
     # conv1d) is the fp32/bf16 floor, gated only against a gross spike.
     from models.tt_dit.tests.models.ltx.test_audio_ltx import _build_torch_stage_c_real
 
-    z = pipeline.tt_audio_decoder.z_channels
+    z = pipeline.tt_mel_decoder.z_channels
     audio_spatial = latent.reshape(1, latent.shape[1], z, latent.shape[2] // z).permute(0, 2, 1, 3).float()
-    mel = pipeline.tt_audio_decoder(audio_spatial)  # TT mel, fed to both TT and torch
+    mel = pipeline.tt_mel_decoder(audio_spatial)  # TT mel, fed to both TT and torch
     w_conv = pipeline.tt_vocoder_with_bwe(mel).squeeze(0).float()
     with torch.no_grad():
         w_torch = _build_torch_stage_c_real(ckpt)(mel.float()).squeeze(0).float()
